correct the type of the signal and of the data, the new signal used from KWindowSystem is using const unsigned long instead of unsigned int
I apologize for the previous bad commit.
Now it is working correctly again.
Lesson learned: always test before commiting.

svn path=/trunk/KDE/kdebase/workspace/; revision=1121254

+++ libs/taskmanager/task.h
@@ -78,11 +78,11 @@ public:
         /**
          * This is corresponding to NET::Property enum reported properties.
          */
-        unsigned int netWindowInfoProperties;
+        unsigned long netWindowInfoProperties;
         /**
          * This is corresponding to NET::Property2 enum reported properties.
          */
-        unsigned int netWindowInfoProperties2;
+        unsigned long netWindowInfoProperties2;
 
         WindowProperties(unsigned int netWinInfoProperties, unsigned int netWinInfoProperties2);
     };
